mind_share/requirements.txt
accelerate==1.3.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==25.1.0
backoff==2.2.1
backports.tarfile==1.2.0
base58==2.1.1
boto3==1.36.13
boto3-stubs==1.36.13
botocore==1.36.13
botocore-stubs==1.36.12
build==1.2.2.post1
bytecode==0.16.1
CacheControl==0.14.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
coincurve==20.0.0
crashtest==0.4.1
cryptography==43.0.3
cytoolz==1.0.1
datasets==2.21.0
ddtrace==2.21.0
Deprecated==1.2.18
dill==0.3.8
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
dulwich==0.22.7
ecdsa==0.19.0
ed25519==1.5
email_validator==2.2.0
envier==0.6.1
eth-hash==0.7.1
eth-keys==0.6.1
eth-typing==5.2.0
eth-utils==5.2.0
exceptiongroup==1.2.2
fastapi==0.111.1
fastapi-cli==0.0.7
fastjsonschema==2.21.1
filelock==3.17.0
fire==0.6.0
fireworks-ai==0.15.12
frozenlist==1.5.0
fsspec==2024.6.1
ftfy==6.3.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.26.0
httpx-sse==0.4.0
httpx-ws==0.7.1
huggingface-hub==0.28.1
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==25.6.0
langchain==0.3.17
langchain-core==0.3.33
langchain-openai==0.3.3
langchain-text-splitters==0.3.5
langsmith==0.3.5
litellm==1.63.11
loguru==0.7.3
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy-boto3==1.35.70
mypy-boto3-s3==1.36.9
near-api-py==0.1.0
nearai==0.1.16
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.66.3
opentelemetry-api==1.31.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pandas-stubs==2.2.2.240807
pathspec==0.12.1
peft==0.10.0
pillow==11.1.0
pkginfo==1.12.0
platformdirs==4.3.6
pluggy==1.5.0
poetry==2.0.1
poetry-core==2.0.1
propcache==0.2.1
protobuf==5.29.3
psutil==5.9.8
py-near==1.1.50
py_near_primitives==0.2.3
pyarrow==19.0.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyMySQL==1.1.1
PyNaCl==1.5.0
pyproject_hooks==1.2.0
pytest==8.3.4
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
RapidFuzz==3.12.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.22.3
s3transfer==0.11.2
safetensors==0.5.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.37.2
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
tensorboardX==2.6.2.2
termcolor==2.5.0
tiktoken==0.8.0
tokenizers==0.21.0
tomli==2.2.1
tomlkit==0.13.2
toolz==1.0.0
torch==2.2.2
tqdm==4.67.1
transformers==4.48.2
trove-classifiers==2025.1.15.22
tweepy==4.15.0
typer==0.12.5
types-awscrt==0.23.9
types-cffi==1.16.0.20241221
types-colorama==0.4.15.20240311
types-docutils==0.21.0.20241128
types-psutil==6.1.0.20241221
types-Pygments==2.19.0.20250107
types-pyOpenSSL==24.1.0.20240722
types-python-dateutil==2.9.0.20241206
types-pytz==2025.1.0.20250204
types-PyYAML==6.0.12.20241230
types-redis==4.6.0.20241004
types-regex==2024.11.6.20241221
types-s3transfer==0.11.2
types-setuptools==71.1.0.20240818
types-ujson==5.10.0.20240515
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
uvicorn==0.32.1
uvloop==0.21.0
virtualenv==20.29.1
watchfiles==1.0.4
wcwidth==0.2.13
websockets==14.2
wrapt==1.17.2
wsproto==1.2.0
xattr==1.1.4
xmltodict==0.14.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
